Merge array initialise and array empty
The EmptyArray() function is very similar to the array initialisation
code in the Array class constructor. Sets the value to blank the other
creates a new object with a blank string.

Should combine these to create new object if element is Null and and set
value to blank if object exists.